// Reconcilify/src/app/leavers/dashboard/models/leavers-kpi.interface.ts
export interface ILeaversKpis{
    acknowledgementRate: number;
    pendingAcknowledgementTasks: number;
    previousDayLeavers: number;
    currentWeekLeavers: number;
    currentMonthLeavers: number;
    yearToDateLeavers: number;
    totalAcknowledgementTasks: number;
    currentWeekBackdatedLeavers: number;
}